+++ font.h
@@ -11,7 +11,7 @@ void font_init(int scaling);
 void font_fillchar(uint32_t * dst_pointer, int dst_char_x, int dst_char_y,
 		   int32_t pitch, uint32_t front_color, uint32_t back_color);
 void font_render(uint32_t * dst_pointer, int dst_char_x, int dst_char_y,
-		 int32_t pitch, uint8_t ch, uint32_t front_color,
+		 int32_t pitch, uint32_t ch, uint32_t front_color,
 		 uint32_t back_color);
 void font_get_size(uint32_t * char_width, uint32_t * char_height);
